% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/alimintest_data.r
\docType{data}
\name{alimintestData}
\alias{alimintestData}
\title{AlimIntest data}
\format{A list of 4 data frames}
\usage{
alimintestData
}
\description{
This dataset contains all metadata and microbial data from the
AlimIntest project.
}
\section{Variables}{


\itemize{
\item \code{metadata}: study design with SCFA, qPCR and comet assays
			\itemize{
   \item idx:	INDEX
   \item Sample_ID; Sample_ID2:	Sample identifications
   \item Time point:	Time point corresponding to study design illustrated on Fig. S1
   \item Subject_Id:	Subject identification
   \item Diet_run:	Diet run "r10_40" correspond to 10g fiber per day between time point n째2 and time point n째3 and 40g fiber per day between time point n째4 and time point n째5; "r40_10" correspond to the cross over to "r10_40". See Fig. S1
   \item Seq_filtered:	Sequence after quality checking from LOTUS pipeline
   \item acetate; propionate; isobutyrate; butyrate; isovalerate; valerate; isocaproate; caproate:	Short Chain Fatty Acid measurement  (mM) by gas-liquid chromatography
   \item log_all_Bacteria; log_Cleptum_group; log_Bcoccoides_group; log_Bacteroides_Prevotella; log_Bifidobacteria; log_Ecoli; log_Lactobacillus:	Quantitative PCR assays expressed in Log10 of number of equivalent bacterial cells
   \item fecal_water_pH:	Fecal water pH measurement
   \item Comet_assay:	Single Cell Gel Electrophoresis assay expressed in % of DNA in the tail
			}
\item \code{otu}: Microbial OTU table
\item \code{tax}: OTU Greengenes taxonomy
\item \code{basalqpcr}: extra qpcr data which correspond to time point 1 and time point 6 (see Fig. S1)
}
}
\examples{
head(alimintestData$metadata)
}

