% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/alimintest_data.r
\docType{data}
\name{basal_diet}
\alias{basal_diet}
\title{Basal diet AlimIntest data}
\format{a data frame with 114 rows and 127 columns}
\usage{
basal_diet
}
\description{
This dataset contains basal diet from individual before and after
inclusion into the AlimIntest study
}
\section{Variables}{


\itemize{
\item \code{subject}: Subject identification
\item \code{visit}: before or after nutritional intervention
\item \code{day}: nb of days before or after nutritional intervention
\item \code{food}: 124 food items, "1" means food item included into diet by the subject
}
}
\examples{
head(basal_diet)
}

