% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/alimintest_data.r
\name{calculate_delta}
\alias{calculate_delta}
\title{calculate delta for a variable}
\usage{
calculate_delta(Subject_Id, time.point, Diet_run, metadata,
  var_name = "variable")
}
\arguments{
\item{Subject_Id}{Subject_Id should be a factor}

\item{time.point}{time.point should be a factor with levels 2,3,4,5}

\item{Diet_run}{should be a factor with levels r40-10 and r10-40}

\item{metadata}{data to be processed}

\item{var_name}{the variable name to be processed inside metadata}
}
\value{
res result
}
\description{
This functions permit to calculate difference before and after diet
}
\examples{
data(alimintestData)
calculate_delta(Subject_Id=alimintestData$metadata$Subject_Id,time.point=alimintestData$metadata$Time.point,
               Diet_run=alimintestData$metadata$Diet_run,metadata=alimintestData$metadata,var_name="acetate")
}

