% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/alimintest_data.r
\docType{data}
\name{metatrans_alimintest}
\alias{metatrans_alimintest}
\title{Metatranscriptomics AlimIntest data}
\format{A list of 4 data frames}
\usage{
metatrans_alimintest
}
\description{
This dataset contains Metatranscriptomics data from the
AlimIntest project.
}
\section{Variables}{


\itemize{
\item \code{tax_metatrans_summary}: taxonomic table from metatranscriptomics sequencing
\item \code{tax_16S_summary}: taxonomic table from 16S rRNA genes reads
\item \code{SubCatKegg}: KEGG categories from metatranscriptomics sequencing
\item \code{metadata}: minimal set of metadata for each sample analysed by metatranscriptomics
}
}
\examples{
head(metatrans_alimintest$metadata)
}

