% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/alimintest_data.r
\docType{data}
\name{replicat_454}
\alias{replicat_454}
\title{454 sequencing data with 4 replicate samples from two manufacturers}
\format{1 data frame with 87782 rows (reads) and 4 columns (variables)}
\usage{
replicat_454
}
\description{
This dataset contains  Sanger VS 454 sequencing data from the same DNA sample
AlimIntest project.
}
\section{Variables}{


\itemize{
\item \code{samples}: sample id
\item \code{manufacturer}: sequencing 454 manufacturer
\item \code{tax}: taxonomy assigned by RDP classifier at genus levels
\item \code{conf}: confidence score from the RDP classifier.
}
}
\examples{
head(replicat_454)
}

