% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{sanger_454}
\alias{sanger_454}
\title{AlimIntest Sanger VS 454 sequencing data}
\format{1 data frame with 15504 rows (reads) and 5 columns (variables)}
\usage{
sanger_454
}
\description{
This dataset contains  Sanger VS 454 sequencing data from the same DNA sample
AlimIntest project.
}
\section{Variables}{


\itemize{
\item \code{samples}: sample replicates id
\item \code{technology}: sequencing technology
\item \code{run}: how replicates were sequenced
\item \code{tax}: taxonomy assigned by RDP classifier at genus levels
\item \code{conf}: confidence score from the RDP classifier.
}
}
\examples{
head(sanger_454)
}

